+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 - Add: `resample_dataframe` utility function. Useful for resampling dataframes with multiple columns that are not candlestick data 
 - Add: `fix_bad_wicks(bad_open_close_threshold)` to massage the data - There are ~60 broken open price data points for U/ni v3 price feeds, cause unknown
 - Add: `filter_for_blacklisted_tokens`
+- Add: `DEXPair.from_series` helper
 - 
 # 0.22.14